fix(Line) 메소드 분리
